requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
mysqlclient==2.2.0
python-multipart==0.0.6
redis==5.0.1
openai==1.3.7
anthropic>=0.18.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
httpx==0.25.2
asyncio-throttle==1.0.2 